Break out sorted int set logic to own class
